backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deepagents>=0.3.0",
    "fastapi>=0.124.4",
    "langchain>=1.2.0",
    "langchain-google-genai>=4.0.0",
    "langgraph-cli[inmem]>=0.4.10",
    "markdownify>=1.2.2",
    "prompt-toolkit>=3.0.52",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "watchdog>=6.0.0",
]
